% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gios_download.R
\name{gios_download}
\alias{gios_download}
\title{Pobieranie plikowej bazy danych jakosci powietrza}
\usage{
gios_download(url, rok = "2000", path = "", ...)
}
\arguments{
\item{url}{charactet, adres pliku archiwum, dostępny jest w obiekcie
`zrodlo`.}

\item{rok}{character, zawiera informacje o roku z którego chcemy pozyskać
dane np. \code{rok = "2000"}.}

\item{path}{character, scieżka dostępu do przestrzeni w której chcemy
utowrzyć plikową baze danych, path = "" oznacza, że dane zostaną zapisane w
folderze roboczym.}

\item{...}{patrz argumenty funkcji \code{\link[utils]{download.file}}, w
systemie windows ustawić \code{mode = "wb"}}
}
\value{
`list` z nazwami plików, które zostały pobrane do lokalizacji `path`.
}
\description{
Funkcja pobiera pliki archiwum z poratalu
\url{http://powietrze.gios.gov.pl/pjp/archives}. Rozpakowywuje do wskazanej
lokalizacji. Zwraca listę plików, które są dostępne w plikowej bazie danych.
Lista plików może posłużyć jak obiekt wejściowy dla funkcji wczytywania
danych \code{gios_read}
}
\examples{
# Pobranie danych za rok 2000 do folderu projektu.
gios_downland(url = zrodlo[1,1],
              rok = "2000",
              path = "",
              mode = "wb")
}
