% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gios_kody.R
\name{gios_kody}
\alias{gios_kody}
\title{Korekta starych nazw stacji}
\usage{
gios_kody(data, meta)
}
\arguments{
\item{data}{data.frame wczytany za pomoca funkcji \code{\link{gios_read}} z
nieuporzadkowanymi oznaczeniami kodow stacji.}

\item{meta}{data.frame pozyskany za pomoca funkcji \code{\link{gios_metadane}} zawiera
informacje o starych i aktulanych kodach stacji.}
}
\value{
zwraca obiekt `data` z zmodyfikowaną zmienna `kod``
}
\description{
Funkcja identyfikuje stare kody stacji jakosci powietrza w wczytanych danych
- \code{\link{gios_read}}. Nastepnie zamienia je na nowe kody stacji.
wymaga pozyskania również metadanych - \code{\link{gios_metadane}}.
}
\examples{
## Pobranie plikowej bazy danych na dysk z <powietrze.gios.gov.pl>
pliki <- gios_download(url = zrodlo[1,1],
                       rok = zrodlo[1,2],
                       path = "",
                       mode = "wb")

# Wczytanie danych o stÄ™ĹĽeniach Ĺ›redniodobowych PM10
pm10 <- gios_read(nazwa = pliki[5],
                  czas_mu = "24h",
                  path =  "")

# pozyskanie metadanych
metadane  <- gios_metadane(type = "meta",
                           download = T,
                           path = "",
                           mode = "wb")

# kerekta zmienej kod
pm10_new <- gios_kody(data = pm10, meta = metadane)

# porownanie starych i nowych kodow stacji
data.frame(new  = flatten_chr(unique(pm10_new[1])),
           old  = flatten_chr(unique(pm10[1])),
           test = (unique(pm10[1]) == unique(pm10_new[1])))
}
