% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gios_metadane.R
\name{gios_metadane}
\alias{gios_metadane}
\title{Funkcja pobierania i wczytywania metadanych}
\usage{
gios_metadane(type = "stacje", download = TRUE, path = "", mode = "wb")
}
\arguments{
\item{type}{charcter, definiuje jakie metadane zostaną wczytane. dostępne są
trzy argumenty \code{type = "stacje"} - metadane stacji, \code{type =
"stanowiska"} - charakterystyka stanowisk pomiarowych, \code{type = "statystyki"} -
statystyki opisowe wyznaczone dla każdej substanji.}

\item{download}{logical, jeśli \code{TRUE} pobierz dane. Jeśli uruchamiasz
funkcje po raz drugi, wybierz \code{FALSE} oraz wskaż ścieżkę dostępu do
pliku.}

\item{path}{character, scieżka dostępu do przestrzeni w której chcemy
utowrzyć plikową baze danych, path = "" oznacza, że dane zostaną zapisane w
folderze roboczym.}
}
\value{
\code{\link[tibble]{tibble}}
}
\description{
Funkcja pobiera trzy rodzaje metadanych i wczytuje je jako obiekt
\code{\link[tibble]{tibble}}. Wczytuje metadane stacji jakości,
Szczegółowe informacje o stanowiskach pomiarowych oraz podstawowe statystyki
opisowe danych pomiarowych, oddzielnie dla każdego zanieczyszczenia powietrza,
substancji i roku kalendarzowego. Ostatni obiekt jest listą złożoną z
obiektów \code{\link[tibble]{tibble}}, które mają różną ilość kolumn,
ponieważ dla każdej substancji liczone są inne staystyki opisowe.
}
\examples{
# Pobranie pliku metadanych stacji jakoĹ›ci powietrza do katalogu projektu.
meatadane <- gios_metadane(type = "stacje",
                           download = T,
                           path = "",
                           mode = "wb")
}
