% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gios_read.R
\name{gios_read}
\alias{gios_read}
\title{Wczytywanie danych pomiarowych}
\usage{
gios_read(nazwa, czas_mu = "24g", path = "")
}
\arguments{
\item{nazwa}{- a character skalar, nazwa pliku który chcemy wczytać.}

\item{czas_mu}{a charcter skalar, można stosować tylko \code{czas_mu = "1g"}
dla średnich 1-h lub \code{czas_mu = "24g"}}

\item{path}{character, scieżka dostępu do przestrzeni w której chcemy
utowrzyć plikową baze danych, path = "" oznacza, że dane zostaną zapisane w
folderze roboczym.}
}
\value{
\code{\link[tibble]{tibble}}
}
\description{
Funkcja wczytuje dane pomiarowe z pobranej plikowej bazy danych. W plikowej
bazie danych dostępne są dwa różne zestawy danych. tj. dane średnie 1-h i
24-h. Jednorazowo pobierany jest pełny zestaw danych pomiarowych dla całej
polski za dany rok kalendarzowy.
}
\examples{
#' ## Pobranie plikowej bazy danych na dysk z <powietrze.gios.gov.pl>
pliki <- gios_download(url = zrodlo[1,1],
                       rok = zrodlo[1,2],
                       path = "",
                       mode = "wb")

# Wczytanie danych o stężeniach średniodobowych PM10
pm10 <- gios_read(nazwa = pliki[5],
                  czas_mu = "24h",
                  path =  "")
}
