% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gios_vis.R
\name{gios_vis}
\alias{gios_vis}
\title{Mapa lokalizacji stacji jakości powietrza}
\usage{
gios_vis(data = meta)
}
\arguments{
\item{data}{- \code{\link[tibble]{tibble}}, obiektem wejsciowym jest
obiekt generowany przez funkcję \code{\link{gios_read}}.}
}
\value{
\code{\link[leaflet]{leaflet}}
}
\description{
Funkcja tworzy interaktywną mapę lokalizacji stacji oraz przedstawia rodzaj
stacji ze wzgledu na źródło odziaływania. Wyświetla kod stacji, co pozwala w
łatwy sposób identyfikować intersujące nas dane pomiarowe.
}
\examples{
# Pobranie pliku metadanych stacji jakości powietrza do katalogu projektu.
meta <- gios_metadane(type = "meta",
                      download = T,
                      path = "",
                      metod = "wb")
# wyświetlenie metadanych
gios_vis(data = meta \%>\% filter(status == "aktywny"))
}
