% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giosimport.R
\docType{package}
\name{giosimport}
\alias{giosimport}
\title{giosimport: Pakiet pobierania i wczytywania danych o jakosci powietrza}
\description{
Pakiet składa się z 5 funkcji, do pobierania metadanych i danych, wczytywania
danych do R oraz wstepnej wizualizacji pozwalajacej zapoznać się z danymi.
Pobierane są dane udostepniane za posrednictwem
\url{http://powietrze.gios.gov.pl/pjp/archives}
}
\section{UWAGA}{
 Funkcje zawarte w pakiecie tworzą plikową bazę danych na
  dysku lokalnym. Cała baza danych możej zajmować prawie 700 MB. Istnieje
  możliwość pobierania tylko wybranych części plikowej bazy danych. W tym
  przypadku dane są pobierane dla każdego roku oddzielnie. Dostępne dane nie
  są aktualizowane, więc wystarczy pobrać dane tylko raz. Warto umieścić je w
  odpowiedniej lokalizacji.
}

