% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zrodlo}
\alias{zrodlo}
\title{Adresy url do plikowej bazy danych GIOS}
\format{
Ramka danych zawiera:
\describe{
 \item{url}{adresy url}
 \item{rok}{rok, którego dotyczy zestaw danych od 2010 - 2020}
}
}
\source{
Główny Inspektorat Ochrony Środowiska. \url{http://powietrze.gios.gov.pl/pjp/archives}
}
\usage{
zrodlo
}
\description{
Adresy url do plikowej bazy danych GIOS
}
\examples{
\dontrun{
 zrodlo
}
}
\keyword{datasets}
